+++ lib/generators/solid_objects/templates/solid_objects.rb
@@ -7,6 +7,7 @@ SolidObjects.configure do |configuration|
   configuration.reminder_scheduler_count = 1
   configuration.authorize_message = ->(**) { false }
   configuration.authorize_query = ->(**) { false }
+  configuration.authorize_destroy = ->(**) { false }
   configuration.authorize_subscription = ->(**) { false }
   configuration.authorize_administration = ->(**) { false }
 end